requirements.txt
FastApi==0.85.0
uvicorn==0.18.3
SQLAlchemy==1.4.3
pickleDB==0.9.2
aiosqlite==0.17.0
python-dotenv==0.21.0